pyproject.toml
[project]
name = "ocr-formatter"
version = "0.1.0"
description = "An application that extracts text from images using OCR and formats it using LLM"
authors = [
    {name = "javierabao"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytesseract>=0.3.10",
    "pillow>=10.1.0",
    "python-docx>=1.0.0",
    "markdown>=3.5",
    "typer>=0.9.0",
    "rich>=13.6.0",
    "transformers>=4.34.0",
    "torch>=2.1.0",
    "accelerate>=0.24.0",
    "sentencepiece>=0.1.99",
    "tiktoken>=0.5.1",
    "einops>=0.7.0",
    "fastapi>=0.92.0",
    "uvicorn[standard]>=0.22.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.1.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
